Fix: 계곡 정보 수정 및 새로운 계곡 등록 시 '물놀이구역' 필드 타입 변경(Double -> Integer)

+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/form/water_place/WaterPlaceForm.java
@@ -40,7 +40,7 @@ public class WaterPlaceForm {
     @Pattern(regexp = "(일반지역|중점관리지역|위험지역)$")
     private String managementType; // 관리유형(일반지역, 중점관리지역)
     @Min(0)
-    private Double waterPlaceSegment; // 물놀이구간(m)
+    private Integer waterPlaceSegment; // 물놀이구간(m)
     @NotNull
     @Min(0)
     private Double deepestDepth; // 수심(깊은곳)
@@ -88,7 +88,7 @@ public class WaterPlaceForm {
         this.address = waterPlace.getAddress();
         this.waterPlaceCategory = waterPlace.getWaterPlaceCategory();
         this.managementType = waterPlace.getManagementType();
-        this.waterPlaceSegment = waterPlaceSegment.equals("-") ? null : Double.valueOf(waterPlaceSegment);
+        this.waterPlaceSegment = waterPlaceSegment.equals("-") ? null : Integer.valueOf(waterPlaceSegment);
         this.deepestDepth = Double.valueOf(deepestDepth);
         this.avgDepth = Double.valueOf(waterPlaceDetail.getAvgDepth());
         this.annualVisitors = annualVisitors.isEmpty() ? null : Double.valueOf(annualVisitors);
